+++ common/header-didattica.php
@@ -17,6 +17,7 @@
 		<ul class="nav navbar-nav top-navbar-nav">
 		<?php if(getSettingsValue('config','sportelli', false)) : ?>
 			<a href="<?php echo $__application_base_path; ?>/didattica/sportello.php" class="btn btn-default navbar-btn btn-orange4" role="button"><span class="glyphicon glyphicon-object-align-horizontal"></span>&ensp;Sportelli </a>
+			<a href="<?php echo $__application_base_path; ?>/didattica/studente.php" class="btn btn-default navbar-btn btn-lima4" role="button"><span class="glyphicon glyphicon-user"></span>&ensp;Studenti </a>
 		<?php endif; ?>
 		</ul>