+++ changelog.md
@@ -1,3 +1,10 @@
+## Version 1.1.3
+
+##### Improvements
+- segreteria: import docenti da file excel
+- ore attirbuite separate, inserite in fatte e previste, con aggiornamento possibile da dirigente
+- didattica: list, gestione studenti e import da file excel
+
 ## Version 1.1.2
 
 ##### Improvements